pyproject.toml
[project]
name = "scalping-bot"
version = "0.1.0"
description = "Automated spot scalping bot for OKX (BTC/ETH/SOL/BNB) with multi-timeframe signals and DCA grid"
requires-python = ">=3.11"
dependencies = [
    "python-okx>=0.3.5",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "rich>=13.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scalping_bot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
